requirements.txt
allure-python-commons~=2.15.0
selene~=2.0.0rc9
pytest~=8.4.1